+++ Examples/SynthTest/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("SynthTest")]
+[assembly: AssemblyDescription("")]
+[assembly: Guid("9DDF90E4-6F4E-42AB-BE25-E210440CAA0D")]