+++ Framework/Audio/Synthesizer/Piano/PianoComponent.cs
@@ -7,6 +7,8 @@
     using System.Linq;
 
     public class PianoComponent : BaseComponent, IUpdateableComponent {
+        private readonly Sprite _blackKeyPressed;
+        private readonly Sprite _blackKeyUnpressed;
 
         private readonly FilterSortCollection<IClickablePianoComponent> _clickables = new FilterSortCollection<IClickablePianoComponent>(
             c => c.IsClickable,
@@ -16,22 +18,54 @@
 
         private readonly NotifyCollectionChangedEventHandler _componentChildrenChangedHandler;
         private readonly LiveSongPlayer _songPlayer;
+        private readonly Sprite _whiteKeyPressed;
+        private readonly Sprite _whiteKeyUnpressed;
         private Camera _camera;
-
         private IClickablePianoComponent _currentClickable;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="PianoComponent"/> class.
         /// </summary>
         /// <param name="songPlayer">The song player.</param>
-        public PianoComponent(LiveSongPlayer songPlayer) : base() {
-            this._songPlayer = songPlayer;
+        /// <param name="whiteKeyUnpressed">The white key unpressed sprite.</param>
+        /// <param name="whiteKeyPressed">The white key pressed sprite.</param>
+        /// <param name="blackKeyUnpressed">The black key unpressed sprite.</param>
+        /// <param name="blackKeyPressed">The black key pressed sprite.</param>
+        public PianoComponent(LiveSongPlayer songPlayer, Sprite whiteKeyUnpressed, Sprite whiteKeyPressed, Sprite blackKeyUnpressed, Sprite blackKeyPressed) : base() {
+            this._songPlayer = songPlayer ?? throw new ArgumentNullException(nameof(songPlayer));
+            this._whiteKeyUnpressed = whiteKeyUnpressed ?? throw new ArgumentNullException(nameof(whiteKeyUnpressed));
+            this._whiteKeyPressed = whiteKeyPressed ?? throw new ArgumentNullException(nameof(whiteKeyPressed));
+            this._blackKeyUnpressed = blackKeyUnpressed ?? throw new ArgumentNullException(nameof(blackKeyUnpressed));
+            this._blackKeyPressed = blackKeyPressed ?? throw new ArgumentNullException(nameof(blackKeyPressed));
             this._componentChildrenChangedHandler = new NotifyCollectionChangedEventHandler(this.Component_ChildrenChanged);
         }
 
         /// inheritdoc />
         public void Update(FrameTime frameTime, InputState inputState) {
-            throw new NotImplementedException();
+            var mouseWorldPosition = this._camera.ConvertPointFromScreenSpaceToWorldSpace(inputState.CurrentMouseState.Position);
+            var isHandled = false;
+            if (this._currentClickable != null) {
+                if (inputState.CurrentMouseState.LeftButton == ButtonState.Pressed) {
+                    isHandled = this._currentClickable.TryHoldClick(mouseWorldPosition);
+                }
+                else {
+                    this._currentClickable.EndClick();
+                    this._currentClickable = null;
+                    isHandled = true;
+                }
+            }
+
+            if (!isHandled && inputState.CurrentMouseState.LeftButton == ButtonState.Pressed) {
+                this._clickables.RebuildCache();
+                foreach (var clickable in this._clickables) {
+                    if (clickable.TryClick(mouseWorldPosition)) {
+                        this._currentClickable = clickable;
+                        break;
+                    }
+                }
+            }
+
+            this._songPlayer.Buffer(0.8f);
         }
 
         protected override void Initialize() {
@@ -40,7 +74,12 @@
 
             foreach (var pitch in pitches) {
                 foreach (var note in notes) {
-                    this.AddChild(new PianoKeyComponent(note, pitch, this._songPlayer));
+                    if (note.IsNatural()) {
+                        this.AddChild(new PianoKeyComponent(note, pitch, this._songPlayer, this._whiteKeyPressed, this._whiteKeyUnpressed));
+                    }
+                    else {
+                        this.AddChild(new PianoKeyComponent(note, pitch, this._songPlayer, this._blackKeyPressed, this._blackKeyUnpressed));
+                    }
                 }
             }
 
@@ -53,8 +92,6 @@
             this.SubscribeToChildrenChanged(this._componentChildrenChangedHandler);
         }
 
-        //// public void Update(FrameTime frameTime) => this.Update(frameTime, MonoGameMouse.Instance.GetState(), MonoGameKeyboard.Instance.GetState());
-
         private void Component_ChildrenChanged(object sender, NotifyCollectionChangedEventArgs e) {
             if (e.Action == NotifyCollectionChangedAction.Add) {
                 foreach (var component in e.NewItems.OfType<IClickablePianoComponent>()) {
@@ -66,33 +103,6 @@
                     this._clickables.Remove(component);
                 }
             }
-        }
-
-        private void Update(MouseState mouseState) {
-            var mouseWorldPosition = this._camera.ConvertPointFromScreenSpaceToWorldSpace(mouseState.Position);
-            var isHandled = false;
-            if (this._currentClickable != null) {
-                if (mouseState.LeftButton == ButtonState.Pressed) {
-                    isHandled = this._currentClickable.TryHoldClick(mouseWorldPosition);
-                }
-                else {
-                    this._currentClickable.EndClick();
-                    this._currentClickable = null;
-                    isHandled = true;
-                }
-            }
-
-            if (!isHandled && mouseState.LeftButton == ButtonState.Pressed) {
-                this._clickables.RebuildCache();
-                foreach (var clickable in this._clickables) {
-                    if (clickable.TryClick(mouseWorldPosition)) {
-                        this._currentClickable = clickable;
-                        break;
-                    }
-                }
-            }
-
-            this._songPlayer.Buffer(0.8f);
         }
     }
 }